pyproject.toml
[project]
name = "hn-digest"
version = "0.1.0"
description = "AI-powered daily Hacker News digest in Chinese"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"
